tf.record_as_input/requirements.txt
tensorflow==1.14.0
google-cloud-storage